getting rid of types.h
