+++ libfreerdp-core/ntlmssp.c
@@ -17,8 +17,6 @@
    limitations under the License.
 */
 
-#include "types.h"
-
 #include <time.h>
 #include <openssl/des.h>
 #include <openssl/md4.h>
